vote-app/charts/frontend/app-django-elastic/requirements.txt
django
django-structlog
opentracing_instrumentation
uwsgi 
elastic-apm[flask] 
elastic-apm[opentracing]
structlog